Moved Form into modal and fixed error messages
